# Conflicts:
#	salt/cloud/__init__.py

+++ salt/modules/parted.py
@@ -49,13 +49,17 @@ def __virtual__():
     These are usually provided by the ``parted`` and ``util-linux`` packages.
     '''
     if salt.utils.is_windows():
-        return False
+        return (False, 'The parted execution module failed to load '
+                'Windows systems are not supported.')
     if not salt.utils.which('parted'):
-        return False
+        return (False, 'The parted execution module failed to load '
+                'parted binary is not in the path.')
     if not salt.utils.which('lsblk'):
-        return False
+        return (False, 'The parted execution module failed to load '
+                'lsblk binary is not in the path.')
     if not salt.utils.which('partprobe'):
-        return False
+        return (False, 'The parted execution module failed to load '
+                'partprobe binary is not in the path.')
     return __virtualname__
 
 
@@ -123,8 +127,6 @@ def probe(*devices):
 
 def list_(device, unit=None):
     '''
-    partition.list device unit
-
     Prints partition information of given <device>
 
     CLI Examples:
@@ -191,8 +193,6 @@ def list_(device, unit=None):
 
 def align_check(device, part_type, partition):
     '''
-    partition.align_check device part_type partition
-
     Check if partition satisfies the alignment constraint of part_type.
     Type must be "minimal" or "optimal".
 
@@ -225,8 +225,6 @@ def align_check(device, part_type, partition):
 
 def check(device, minor):
     '''
-    partition.check device minor
-
     Checks if the file system on partition <minor> has any errors.
 
     CLI Example:
@@ -251,11 +249,9 @@ def check(device, minor):
 
 def cp(device, from_minor, to_minor):  # pylint: disable=C0103
     '''
-    partition.check device from_minor to_minor
-
     Copies the file system on the partition <from-minor> to partition
-        <to-minor>, deleting the original contents of the destination
-        partition.
+    <to-minor>, deleting the original contents of the destination
+    partition.
 
     CLI Example:
 
@@ -280,8 +276,6 @@ def cp(device, from_minor, to_minor):  # pylint: disable=C0103
 
 def get_id(device, minor):
     '''
-    partition.get_id
-
     Prints the system ID for the partition. Some typical values are::
 
          b: FAT32 (vfat)
@@ -313,8 +307,6 @@ def get_id(device, minor):
 
 def set_id(device, minor, system_id):
     '''
-    partition.set_id
-
     Sets the system ID for the partition. Some typical values are::
 
          b: FAT32 (vfat)
@@ -372,12 +364,9 @@ def system_types():
 
 def mkfs(device, fs_type):
     '''
-    partition.mkfs device fs_type
-
     Makes a file system <fs_type> on partition <device>, destroying all data
-        that resides on that partition. <fs_type> must be one of "ext2",
-        "fat32", "fat16", "linux-swap" or "reiserfs" (if libreiserfs is
-        installed)
+    that resides on that partition. <fs_type> must be one of "ext2", "fat32",
+    "fat16", "linux-swap" or "reiserfs" (if libreiserfs is installed)
 
     CLI Example:
 
@@ -391,7 +380,11 @@ def mkfs(device, fs_type):
                           'hfs', 'hfs+', 'hfsx', 'NTFS', 'ufs']):
         raise CommandExecutionError('Invalid fs_type passed to partition.mkfs')
 
-    mkfs_cmd = 'mkfs.{0}'.format(fs_type)
+    if fs_type is 'linux-swap':
+        mkfs_cmd = 'mkswap'
+    else:
+        mkfs_cmd = 'mkfs.{0}'.format(fs_type)
+
     if not salt.utils.which(mkfs_cmd):
         return 'Error: {0} is unavailable.'
     cmd = '{0} {1}'.format(mkfs_cmd, device)
@@ -401,9 +394,8 @@ def mkfs(device, fs_type):
 
 def mklabel(device, label_type):
     '''
-    partition.mklabel device label_type
-
     Create a new disklabel (partition table) of label_type.
+
     Type should be one of "aix", "amiga", "bsd", "dvh", "gpt", "loop", "mac",
     "msdos", "pc98", or "sun".
 
@@ -413,26 +405,23 @@ def mklabel(device, label_type):
 
         salt '*' partition.mklabel /dev/sda msdos
     '''
-    _validate_device(device)
-
-    if label_type not in set(['aix', 'amiga', 'bsd', 'dvh', 'gpt', 'loop', 'mac',
-                             'msdos', 'pc98', 'sun']):
+    if label_type not in set([
+        'aix', 'amiga', 'bsd', 'dvh', 'gpt', 'loop', 'mac', 'msdos', 'pc98', 'sun'
+    ]):
         raise CommandExecutionError(
             'Invalid label_type passed to partition.mklabel'
         )
 
-    cmd = 'parted -m -s {0} mklabel {1}'.format(device, label_type)
-    out = __salt__['cmd.run'](cmd).splitlines()
+    cmd = ('parted', '-m', '-s', device, 'mklabel', label_type)
+    out = __salt__['cmd.run'](cmd, python_shell=False).splitlines()
     return out
 
 
 def mkpart(device, part_type, fs_type=None, start=None, end=None):
     '''
-    partition.mkpart device part_type fs_type start end
-
     Make a part_type partition for filesystem fs_type, beginning at start and
-        ending at end (by default in megabytes).  part_type should be one of
-        "primary", "logical", or "extended".
+    ending at end (by default in megabytes).  part_type should be one of
+    "primary", "logical", or "extended".
 
     CLI Examples:
 
@@ -441,13 +430,6 @@ def mkpart(device, part_type, fs_type=None, start=None, end=None):
         salt '*' partition.mkpart /dev/sda primary fs_type=fat32 start=0 end=639
         salt '*' partition.mkpart /dev/sda primary start=0 end=639
     '''
-    _validate_device(device)
-
-    if not start or not end:
-        raise CommandExecutionError(
-            'partition.mkpart requires a start and an end'
-        )
-
     if part_type not in set(['primary', 'logical', 'extended']):
         raise CommandExecutionError(
             'Invalid part_type passed to partition.mkpart'
@@ -459,31 +441,33 @@ def mkpart(device, part_type, fs_type=None, start=None, end=None):
             'Invalid fs_type passed to partition.mkpart'
         )
 
-    _validate_partition_boundary(start)
-    _validate_partition_boundary(end)
+    if start is not None and end is not None:
+        _validate_partition_boundary(start)
+        _validate_partition_boundary(end)
+
+    if start is None:
+        start = ''
+
+    if end is None:
+        end = ''
 
     if fs_type:
-        cmd = 'parted -m -s -- {0} mkpart {1} {2} {3} {4}'.format(
-            device, part_type, fs_type, start, end
-        )
+        cmd = ('parted', '-m', '-s', '--', device, 'mkpart', part_type, fs_type, start, end)
     else:
-        cmd = 'parted -m -s -- {0} mkpart {1} {2} {3}'.format(
-            device, part_type, start, end
-        )
+        cmd = ('parted', '-m', '-s', '--', device, 'mkpart', part_type, start, end)
 
-    out = __salt__['cmd.run'](cmd).splitlines()
+    out = __salt__['cmd.run'](cmd, python_shell=False).splitlines()
     return out
 
 
 def mkpartfs(device, part_type, fs_type, start, end):
     '''
-    partition.mkpartfs device part_type fs_type start end
-
     Make a <part_type> partition with a new filesystem of <fs_type>, beginning
-        at <start> and ending at <end> (by default in megabytes).  <part_type>
-        should be one of "primary", "logical", or "extended". <fs_type> must be
-        one of "ext2", "fat32", "fat16", "linux-swap" or "reiserfs" (if
-        libreiserfs is installed)
+    at <start> and ending at <end> (by default in megabytes).
+
+    <part_type> should be one of "primary", "logical", or "extended". <fs_type>
+    must be one of "ext2", "fat32", "fat16", "linux-swap" or "reiserfs" (if
+    libreiserfs is installed)
 
     CLI Example:
 
@@ -516,10 +500,8 @@ def mkpartfs(device, part_type, fs_type, start, end):
 
 def name(device, partition, name):
     '''
-    partition.name device partition name
-
-    Set the name of partition to name. This option works only on Mac, PC98,
-        and GPT disklabels. The name can be placed in quotes, if necessary.
+    Set the name of partition to name. This option works only on Mac, PC98, and
+    GPT disklabels. The name can be placed in quotes, if necessary.
 
     CLI Example:
 
@@ -550,11 +532,9 @@ def name(device, partition, name):
 
 def rescue(device, start, end):
     '''
-    partition.rescue device start end
-
     Rescue a lost partition that was located somewhere between start and end.
-        If a partition is found, parted will ask if you want to create an
-        entry for it in the partition table.
+    If a partition is found, parted will ask if you want to create an
+    entry for it in the partition table.
 
     CLI Example:
 
@@ -573,13 +553,12 @@ def rescue(device, start, end):
 
 def resize(device, minor, start, end):
     '''
-    partition.resize device minor, start, end
+    Resizes the partition with number <minor>.
 
-    Resizes the partition with number <minor>. The partition will start <start>
-        from the beginning of the disk, and end <end> from the beginning of the
-        disk. resize never changes the minor number. Extended partitions can be
-        resized, so long as the new extended partition completely contains all
-        logical partitions.
+    The partition will start <start> from the beginning of the disk, and end
+    <end> from the beginning of the disk. resize never changes the minor number.
+    Extended partitions can be resized, so long as the new extended partition
+    completely contains all logical partitions.
 
     CLI Example:
 
@@ -609,8 +588,6 @@ def resize(device, minor, start, end):
 
 def rm(device, minor):  # pylint: disable=C0103
     '''
-    partition.rm device minor
-
     Removes the partition with number <minor>.
 
     CLI Example:
@@ -635,11 +612,13 @@ def rm(device, minor):  # pylint: disable=C0103
 
 def set_(device, minor, flag, state):
     '''
-    partition.set device  minor flag state
+    Changes a flag on the partition with number <minor>.
 
-    Changes a flag on the partition with number <minor>. A flag can be either
-        "on" or "off". Some or all of these flags will be available, depending
-        on what disk label you are using.
+    A flag can be either "on" or "off". Some or all of these flags will be
+    available, depending on what disk label you are using.
+
+    Valid flags are: bios_grub, legacy_boot, boot, lba, root, swap, hidden, raid,
+        LVM, PALO, PREP, DIAG
 
     CLI Example:
 
@@ -670,15 +649,14 @@ def set_(device, minor, flag, state):
 
 def toggle(device, partition, flag):
     '''
-    partition.toggle device partition flag
-
-    Toggle the state of <flag> on <partition>
+    Toggle the state of <flag> on <partition>. Valid flags are the same as
+        the set command.
 
     CLI Example:
 
     .. code-block:: bash
 
-        salt '*' partition.name /dev/sda 1 boot
+        salt '*' partition.toggle /dev/sda 1 boot
     '''
     _validate_device(device)
 
@@ -700,8 +678,6 @@ def toggle(device, partition, flag):
 
 def exists(device=''):
     '''
-    partition.exists device
-
     Check to see if the partition exists
 
     CLI Example:
